A new file added - "CommonViews" Common tiltle, game over text, and total score text are moved to the CommonViews file
